+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/preferences/ShortcutsPreferencesController.java
@@ -1,0 +1,22 @@
+package de.mephisto.vpin.ui.preferences;
+
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import javafx.scene.layout.HBox;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.net.URL;
+import java.util.ResourceBundle;
+
+public class ShortcutsPreferencesController implements Initializable {
+  private final static Logger LOG = LoggerFactory.getLogger(ShortcutsPreferencesController.class);
+
+  @FXML
+  private HBox htmlRoot;
+
+  @Override
+  public void initialize(URL url, ResourceBundle resourceBundle) {
+
+  }
+}